Added code for the Arduino
Had to rewrite pretty much the entire script in C, but now I also have an Arduino version, which (as a bonus) includes support for a piezo speaker. Defaults to the built-in LED pin (13) and pin 8 for the speaker - they can work in series with the resistor but it makes the circuit unnecessarily complicated.
No dicts or hashtables in C, so instead we have an ugly switch statement.